pyproject.toml
[tool.poetry]
name = "blastcards"
version = "0.1.0"
description = "A very simple flashcards application to my portfolio"
authors = ["André Carvalho <alac1984@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
uvicorn = {extras = ["standart"], version = "^0.18.3"}
Jinja2 = "^3.1.2"
SQLAlchemy = "^1.4.41"
alembic = "^1.8.1"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.4"
pydantic = {extras = ["email"], version = "^1.10.2"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pynvim = "^0.4.3"
pytest = "^7.1.3"
mypy = "^0.982"
flake8 = "^5.0.4"
black = "^22.10.0"
types-passlib = "^1.7.7"
sqlalchemy2-stubs = "^0.0.2-alpha.29"
requests = "^2.28.1"
Flake8-pyproject = "^1.1.0"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.flake8]
ignore = ['E501', 'W503', 'E203', 'E402', 'E712']
max-line-length = 88
exclude = ['.git', 'migrations', 'db/base.py']

[tool.mypy]
disable_error_code = ['attr-defined', 'import', 'valid-type']

[tool.pytest.ini_options]
markers = ["integration", "unit"]
